revert(listfiles): reverted last changes so branch could be merged to main
