+++ apps/backoffice-v2/src/components/pages/Individual/Individual.page.tsx
@@ -1,12 +1,9 @@
-import { Subject } from 'components/organisms/Subject/Subject';
-import { useIndividual } from 'components/pages/Individual/hooks/useIndividual/useIndividual';
-import { useConsole } from 'hooks/useConsole/useConsole';
+import { Subject } from "components/organisms/Subject/Subject";
+import { useIndividual } from "components/pages/Individual/hooks/useIndividual/useIndividual";
 
 export const Individual = () => {
   const { selectedEndUser, faceAUrl, faceBUrl, info, images, isLoading, whitelist } =
     useIndividual();
-
-  useConsole(images);
 
   // Selected end user
   return (
