+++ services/workflows-service/src/storage/storage.service.ts
@@ -9,20 +9,12 @@ export class StorageService {
 
   async createFileLink({
     fileNameOnDisk,
-    workflowId,
     userId,
-  }: Pick<Prisma.FileCreateInput, 'fileNameOnDisk' | 'userId'> & {
-    workflowId: string;
-  }) {
+  }: Pick<Prisma.FileCreateInput, 'fileNameOnDisk' | 'userId'>) {
     const file = await this.fileRepository.create({
       data: {
         fileNameOnDisk,
         userId,
-        workflowRuntimeData: {
-          connect: {
-            id: workflowId,
-          }
-        }
       },
       select: {
         id: true,
@@ -30,12 +22,6 @@ export class StorageService {
     });
 
     return file.id;
-  }
-
-  async listFileNames({ userId }: IFileIds) {
-    return await this.fileRepository.findManyNames({
-      userId,
-    });
   }
 
   async getFileNameById({ id, userId }: IFileIds) {
